# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FarmFlow API
  description: |
    Egyptian livestock marketplace API.
    All monetary values are in EGP (Egyptian Pound).
    All dates are ISO 8601 strings.
    Image paths are relative to the server root (e.g. `/uploads/filename.jpg`).

    **Authentication:** Pass a JWT token in the `Authorization: Bearer <token>` header.

    **Roles:**
    - `buyer` — can browse, order, review, manage favorites
    - `seller` — can manage listings, animals, dairy, supplies, finances
    - `admin` — full access including approval workflows

    **⚠️ Mobile app endpoint notes:**
    - `GET /api/listings/my` is a dedicated route (seller/admin JWT required) — returns own listings (all statuses)
    - `GET /api/orders/my` is handled by `GET /api/orders` (buyer JWT → own orders only)
    - `PATCH /api/listings/:id/status` does not exist — use `PUT /api/listings/:id` with a `status` body field
    - `PATCH /api/admin/users/:id` does not exist — use `PATCH /api/admin/users/:id/toggle`
  version: 1.0.0
  contact:
    email: a.khalil@soar.inc
servers:
  - url: http://localhost:5001/api
    description: Local development
  - url: https://api.farmflow.app/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MongoId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 64f1a2b3c4d5e6f7a8b9c0d1
...
    Supply:
      type: object
      properties:
        _id:       { $ref: '#/components/schemas/MongoId' }
        seller:    { $ref: '#/components/schemas/MongoId' }
        name:      { type: string }
        category:  { type: string, enum: [feed, veterinary, equipment, seeds, other] }
        description: { type: string }
        quantity:  { type: number }
        unit:      { type: string }
        pricePerUnit: { type: number }
        minOrderQty:  { type: number }
        location:  { type: string }
        deliveryAvailable: { type: boolean }
        deliveryCost: { type: number }
        images:    { type: array, items: { type: string } }
        status:    { type: string, enum: [pending, approved, rejected, sold_out] }
        rejectionReason: { type: string }
        createdAt: { type: string, format: date-time }

    # ── Reviews ─────────────────────────────────────────────────────────────────
...
paths:
...
  /supplies:
    get:
      tags: [Supplies]
      summary: Get supplies (public → approved only; seller JWT → own; admin → all)
      parameters:
        - in: query
          name: category
          schema: { type: string, enum: [feed, veterinary, equipment, seeds, other] }
      responses:
        '200':
          description: Array of supplies
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Supply' }

    post:
      tags: [Supplies]
      summary: Create supply listing (seller only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, category, quantity, unit, pricePerUnit]
              properties:
                name:              { type: string }
                category:          { type: string, enum: [feed, veterinary, equipment, seeds, other] }
                description:       { type: string }
                quantity:          { type: number }
                unit:              { type: string }
                pricePerUnit:      { type: number }
                minOrderQty:       { type: number }
                location:          { type: string }
                deliveryAvailable: { type: boolean }
                deliveryCost:      { type: number }
                images:            { type: array, items: { type: string, format: binary } }
      responses:
        '201':
          description: Supply created (status = pending)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supply' }
  /supplies/{id}:
    get:
      tags: [Supplies]
      summary: Get supply by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Supply detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supply' }

    put:
      tags: [Supplies]
      summary: Update supply (seller owns it or admin)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:              { type: string }
                category:          { type: string, enum: [feed, veterinary, equipment, seeds, other] }
                description:       { type: string }
                quantity:          { type: number }
                unit:              { type: string }
                pricePerUnit:      { type: number }
                minOrderQty:       { type: number }
                location:          { type: string }
                deliveryAvailable: { type: boolean }
                deliveryCost:      { type: number }
                keepImages:        { type: string, description: JSON array of existing paths to keep }
                images:            { type: array, items: { type: string, format: binary } }
      responses:
        '200':
          description: Updated supply
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supply' }

    delete:
      tags: [Supplies]
      summary: Delete supply
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }
  /supplies/{id}/status:
    patch:
      tags: [Supplies]
      summary: Approve, reject or mark sold-out (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:          { type: string, enum: [approved, rejected, sold_out] }
                rejectionReason: { type: string }
      responses:
        '200':
          description: Updated supply
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Supply' }

  # ── REVIEWS ──────────────────────────────────────────────────────────────────
...